requirements.txt
tensorflow==2.1.0
streamlit==0.76.0
numpy==1.19.5
pillow==8.1.0
pandas==1.1.3
re==2.2.1
json==2.0.9
